airControlTower
The code scans two airplanes and their call-sign, distance, direction and altitude. It prints out the distance and difference in height between the two planes and then applies some changes: increase the altitude of Airplane 1 by 4000 feet, decrease the altitude of Airplane 2 by 2000 feet, move Airplane 1 a distance of 10.5 miles on a heading of 50 degrees, move Airplane 2 a distance of 8.0 miles on a heading of 125 degrees. The programs prints out the applied changes and the new positions of the two planes.