Create crackTheCode
The program asks the user to figure out a code. The first number entered has to be 3 , the second number can be 1 or a number between 33 and 100 and the third number has to be positive or divisible by 3 or 7. If the user fails to meet the requirement in any of the requests, then the program exits and he/she has to start over.